* Convert i2d_value from long to long long.

+++ i2d_data.h
@@ -55,7 +55,7 @@ enum i2d_value_type {
 };
 
 struct i2d_value {
-    long value;
+    long long value;
     i2d_string string;
     i2d_string_stack stack;
 };
@@ -73,6 +73,6 @@ typedef struct i2d_value_map i2d_value_map;
 
 int i2d_value_map_init(i2d_value_map **, json_t *, enum i2d_value_type);
 void i2d_value_map_deit(i2d_value_map **);
-int i2d_value_map_get_string(i2d_value_map *, long, i2d_string *);
-int i2d_value_map_get_string_stack(i2d_value_map *, long, i2d_string_stack *);
+int i2d_value_map_get_string(i2d_value_map *, long long, i2d_string *);
+int i2d_value_map_get_string_stack(i2d_value_map *, long long, i2d_string_stack *);
 #endif
